feat(fn_eval): Add fn in fn evaluation
This adds the ability to nest functions and evaluate args before
actually evaluating the main function first. This means changes to how
the parser and evaluator were done. It also lifts restrictions for other
types of ops for length, but since we still only handle ones with two
args the errors still get thrown.